requirements.txt
aws-cdk-lib==2.114.1